requirements.txt
discord.py
python-jobspy
pandas
beautifulsoup4
requests
psycopg2-binary
boto3
requests